+++ src/exceptions/ProviderError.ts
@@ -1,0 +1,6 @@
+export class ProviderError extends Error {
+    constructor(message: string) {
+        super();
+        this.message = "No database provider was specified."
+    }
+}